refactor: :recycle: Move 'validate_file_callback' into playlist module to share with other commands
Add constant SUPPORTED_PLS_FILES with list of playlist file extensions
